Edited function, Removed function after feedback
Removed a function to set the send buffer with a string. This function was removed after feedback due to the fact that this function might not be used at all. Edited a function so it uses the send function instead of the add_send_byte. Added get_send_bytes function which returns the send buffer into a vector.